refactor: change alias path for strategies
