+++ src/types/chat.ts
@@ -1,0 +1,2 @@
+export type Sender = "user" | "assistant";
+export type StrategyType = "transformers" | "pyodide";
